Correct translation of log level into journald priority
The issue is that systemd/journal priority enumeration is exactly the other
way from what the log4perl dispatcher provides. In other words, whatever is
logged as "debug" in log4perl becomes an emergency (and gets broadcast) via
systemd, and whatever is an emergency for log4perl gets swallowed as
unimportant on the system side.

[lkundrak@v3.sk: Taken from an e-mail message, added a comment.]